Move MycoBank & MycoPortal link helpers to new ExternalLinkHelper
